onemoguceno gledanje/brisanje porudzbina koje nisu kreirane od strane ulogovanog korisnika

+++ poslasticarnica/app/Http/Controllers/StavkaKorpeController.php
@@ -18,10 +18,34 @@ class StavkaKorpeController extends Controller
     {
 
        $order = StavkaKorpe::find($id);
-     
+        
        if($order){
+           if($order->user_id==auth()->id()){
             $order->delete();
-            return response()->json("uspesno obrisano!");
+            return response()->json("uspesno obrisano! " );
+           }else{
+            return response()->json("Nemate pravo da obrisete porudzbinu koja nije vasa!" );
+           }
+            
+
+       }else{
+            return response()->json("ne postoji trazeni objekat u bazi!");
+       }
+    }
+
+    public function show($id)  
+    {
+
+       $order = StavkaKorpe::find($id);
+        
+       if($order){
+           if($order->user_id==auth()->id()){
+            
+            return response()->json( $order );
+           }else{
+            return response()->json("Nemate pravo da vidite porudzbinu koja nije vasa!" );
+           }
+            
 
        }else{
             return response()->json("ne postoji trazeni objekat u bazi!");
